Potential fix for repeated homing?
The homed flag is set on the axis motor status, but the params callback
for the controller is used which will be for axis A so there may be some
delay in other axes updating? Bit of a guess, not sure why you would
only see an issue with multiple simultaneous homes though.